Extract add images widget
Also remove custom expandIndexed
