+++ lib/page/marker/add_images_widget.dart
@@ -1,9 +1,9 @@
 import 'dart:io';
 
+import 'package:collection/collection.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:image_picker/image_picker.dart';
-import 'package:insigno_frontend/util/iterable.dart';
 import 'package:insigno_frontend/util/nullable.dart';
 import 'package:insigno_frontend/util/pair.dart';
 import 'package:os_detect/os_detect.dart' as Platform;
